requirements.txt
appnope==0.1.0
asgiref==3.3.0
backcall==0.2.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
decorator==4.4.2
Django==3.1.3
django-cors-headers==3.5.0
Faker==4.14.0
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
mysqlclient==2.0.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
gunicorn==20.0.4
